+++ tests/integration/src/proxy_new.rs
